Recreate LINKS section of CaboVillas.com with Bulma and Vue.
Initial commit after working on for a day or two, even though in a repo in VS
Code... sigh.
Anyway, this page started as a bulma test, and then I "progressivley"
added reactivity with Vue. Got the idea because had added borders to the
"column" divs to show how they changed size responsivly, but it
distracted from the UX, so I added radio buttons to turn them on and off.
Then decided to try my hand at a Select updating dynamic styling, so I
added the functionality to a dropdown in the same section.
Around this point I dropped the class based dynamic styling.
Final touch was to add a watcher (started as computed) to reactively
set the "show border" radio to true when the value for selected changes.